requirements.txt
autopep8==1.5.4
backcall==0.2.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
colorama==0.4.4
decorator==4.4.2
flake8==3.8.4
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
mccabe==0.6.1
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
python-dotenv==0.21.0
pywin32==228
PyYAML==6.0
pyzmq==19.0.2
toml==0.10.1
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wincertstore==0.2
